+++ .gitignore
@@ -3,9 +3,34 @@ __pycache__/
 *.py[cod]
 *$py.class
 
-# C extensions
-*.so
+# Prerequisites
+*.d
+
+# Compiled Object files
+*.slo
+*.lo
 *.o
+*.obj
+
+# Precompiled Headers
+*.gch
+*.pch
+
+# Compiled Dynamic libraries
+*.so
+*.dylib
+*.dll
+
+# Compiled Static libraries
+*.lai
+*.la
+*.a
+*.lib
+
+# Executables
+*.exe
+*.out
+*.app
 
 # Distribution / packaging
 .Python
