+++ cEpiabm/src/dataclasses/household.hpp
@@ -1,0 +1,16 @@
+#ifndef EPIABM_DATACLASSES_HOUSEHOLD_HPP
+#define EPIABM_DATACLASSES_HOUSEHOLD_HPP
+
+namespace epiabm
+{
+
+    class Household
+    {
+    private:
+    public:
+    private:
+    };
+
+} // namespace epiabm
+
+#endif // EPIABM_DATACLASSES_HOUSEHOLD_HPP
